+++ lib/src/block_syntaxes/ordered_list_with_checkbox_syntax.dart
@@ -2,16 +2,9 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '../patterns.dart';
-import 'list_syntax.dart';
+import 'ordered_list_syntax.dart';
 
-/// Parses ordered lists.
-class OrderedListWithCheckBoxSyntax extends ListSyntax {
-  @override
-  RegExp get pattern => olWithCheckBoxPattern;
-
-  @override
-  String get listTag => 'ol_with_checkbox';
-
-  const OrderedListWithCheckBoxSyntax();
+/// Parses ordered lists with checkboxes.
+class OrderedListWithCheckboxSyntax extends OrderedListSyntax {
+  const OrderedListWithCheckboxSyntax();
 }