+++ lib/src/inline_syntaxes/tag_syntax.dart
@@ -6,6 +6,5 @@ import 'delimiter_syntax.dart';
 
 @Deprecated('Use DelimiterSyntax instead')
 class TagSyntax extends DelimiterSyntax {
-  TagSyntax(String pattern, {bool requiresDelimiterRun = false})
-      : super(pattern, requiresDelimiterRun: requiresDelimiterRun);
+  TagSyntax(super.pattern, {super.requiresDelimiterRun});
 }